requirements.txt
pillow==11.1.0
pillow_heif==0.21.0
